requirements.txt
bottle==0.12.23
certifi==2022.12.7
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.4
openai==0.25.0
openpyxl==3.0.10
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
tqdm==4.64.1
twilio==7.16.0
types-pytz==2022.7.0.0
typing-extensions==4.4.0
urllib3==1.26.13
